waste-detection-mlops/requirements.txt
fastapi>=0.111
uvicorn[standard]
streamlit
mlflow
apache-airflow
prometheus-client
folium
streamlit-folium
ultralytics
torch
torchvision
sqlite3
pytest
httpx
python-multipart